Update to VariantSelects so that a size is not automatically selected on the product page.
"More precisely, in our variants.js script, there is a function called this.renderProductInfo() in the method onVariantChange()
What this renderProductInfo() essentially does is it fetches HTML element for <variant-radios> component from the url ${this.dataset.url}?variant=${this.currentVariant.id}&section_id=${this.dataset.section} (this.dataset.url is our store url, and this.currentVariant.id is our variant id) every time we switch a size (the this.currentVariant.id changes every time we change a size, as each size indicates different variant).

Like we discussed multiple times, fetch() function takes a long time to execute, resulting in a dull user experience. There are places where fetch() is a must (e.g. cart, plp filter), whereas in some other places, it should not be used if you want sharper experience.

In our PDP, when we choose a different size, the master variant id and the <variant-radios> component gets updated regardless of the fetch() function.
That being said, fetch() really serves no purpose in the flow - it’s sole objective is to update <variant-radios> component correctly every time we pick a new size, but it gets updated well even without the fetch() function.

Thus, I commented out everything that was inside renderProductInfo(), and only left out this.toggleAndButton(...) function, as this is the only function responsible for updating the atc button status."